requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
contourpy==1.2.0
cryptography==42.0.2
CTkTable==1.1
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
exceptiongroup==1.2.0
fonttools==4.47.0
h11==0.14.0
idna==3.6
importlib-resources==6.1.1
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.4
outcome==1.3.0.post0
packaging==23.2
pandas==1.5.0
Pillow==10.1.0
pycparser==2.21
pyOpenSSL==24.0.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
selenium==4.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
threadpoolctl==3.2.0
trio==0.24.0
trio-websocket==0.11.1
urllib3==1.26.18
urllib3-secure-extra==0.1.0
wsproto==1.2.0
zipp==3.17.0
